rake move_game[g_18_mag]
